+++ web/component/pp-service/src/main/java/com/pp/service/tableFile/TableFileService.java
@@ -3,5 +3,5 @@ package com.pp.service.tableFile;
 public interface TableFileService {
     void addTableFile(String tableName, Long tableId, Integer type, Long fileRecordId);
 
-    void updateTableFile(Long id, Integer flag);
+    void updateTableFile(Long id, byte flag);
 }
